+++ packages/kas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @khanacademy/kas
 
+## 0.2.4
+
+### Patch Changes
+
+-   a15b0e86: Add 'perseus-build-settings' as a dev dep to packages that were missing it
+
 ## 0.2.3
 
 ### Patch Changes