+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @khanacademy/perseus-editor
 
+## 0.3.1
+
+### Patch Changes
+
+-   @khanacademy/perseus@0.2.5
+
+## 0.3.0
+
+### Minor Changes
+
+-   8315e911: Don't re-export Perseus from @khanacademy/perseus-editor
+
+### Patch Changes
+
+-   Updated dependencies [5d86a625]
+    -   @khanacademy/perseus@0.2.4
+
+## 0.2.0
+
+### Minor Changes
+
+-   309cd98d: Content editor components must now specify a previewURL instead of frameSource
+
+### Patch Changes
+
+-   Updated dependencies [73095170]
+    -   @khanacademy/perseus@0.2.3
+
 ## 0.1.3
 
 ### Patch Changes